Bugfixing and simplification
My vectorize function was not checking to see if a process was complete before returning! Wow, what a bugger!

+++ Synthesizer/metadata/stencils.py
@@ -1,10 +1,10 @@
 import json
 import os
+import math
 
 from Raster.Raster import Raster
-from Raster import filters, analyze, math_utilities
-import numpy as np
-import math
+from Raster import filters, analyze
+from Utilities import math as math_util
 
 
 def make_stencil(stencil_name, quantity, stencil_staging, stencil_configs, colorize, relative_path):
@@ -29,12 +29,14 @@ def make_stencil(stencil_name, quantity, stencil_staging, stencil_configs, color
 
 def apply_stencil(data, paths, resourcepack):
     mapping_path, json_data = data
+    print(mapping_path)
 
     image_components = []
     for cluster_data in json_data['segment_dicts']:
         segment = Raster.from_path(
             paths.resource_skeletons + '//' + resourcepack + '//'
             + json_data['group_name'] + '_' + str(cluster_data['id']+1) + '.png', "RGBA")
+
         if not cluster_data['colorize']:
             image_components.append(segment)
             continue
@@ -42,50 +44,25 @@ def apply_stencil(data, paths, resourcepack):
         # Adjust contrast
         contrast_mult = abs(analyze.variance(segment, 'V') - math.sqrt(cluster_data['variance'])) * .1
         segment = filters.contrast(segment, contrast_mult)
-
+        print('Success 0.5')
         # Adjust coloration
-        layer_count = len(cluster_data['hues'])*2
+        layer_count = len(cluster_data['hues'])
+        print(layer_count)
         components = filters.value_decomposite(segment, layer_count)
-
-        # sat_poly_raw = math_utilities.polyfit(
-        #     np.linspace(0, 1, len(cluster_data['sats'])), cluster_data['sats'])
-        target_sat = math_utilities.linear_mean(cluster_data['sats'])
-        target_hue = math_utilities.circular_mean(cluster_data['hues'])
-        # linear_mapped_hues = (np.array(sorted_hues) - sorted_hues[0]) % 1
-        # hue_poly = math_utilities.polyfit(np.linspace(0, 1, len(cluster_data['hues'])), linear_mapped_hues)
-
+        print("Success 1")
         colorized_components = []
-        for index, layer in enumerate(components):
-
-            normalized_index = float(index) / len(components)
-
-            # hue_target = (math_utilities.polysolve(hue_poly, normalized_index) + sorted_hues[0]) % 1
-            # sat_target = math_utilities.polysolve(sat_poly_raw, normalized_index)
-
-            # Reduce sat in lighter areas of image
-            mean_lightness = analyze.mean(layer, 'V')
-            sat_reduction = pow(mean_lightness, len(components)) * target_sat
-            target_sat -= sat_reduction
-            target_sat = math_utilities.clamp(target_sat)
-            print(mapping_path + " sat " + str(target_sat))
-            layer = filters.colorize(layer, target_hue, target_sat, 0, 1., 1.0, 0)
+        for i, layer in enumerate(components):
+            print(i)
+            print(cluster_data['hues'])
+            print(cluster_data['hues'][i])
+            layer = filters.colorize(layer, cluster_data['hues'][i], cluster_data['sats'][i], 0, 1, 1, 0)
             colorized_components.append(layer)
-
+        print("Success 2")
         staged_image = filters.composite(colorized_components)
-
         # Adjust lightness
         lightness_adjustment = cluster_data['lightness'] - analyze.mean(segment, 'V')
-        # print(lightness_adjustment)
-        while(analyze.mean(staged_image, 'V') - abs(lightness_adjustment)
-              < cluster_data['lightness']
-              < analyze.mean(staged_image, 'V') + abs(lightness_adjustment)):
-
-            # print(template_path + '\n'
-            #       + str(cluster_data['lightness']) + ' ' + str(analyze.mean(staged_image, 'V')) + '\n'
-            #       + str(lightness_adjustment))
-
-            staged_image = filters.brightness(staged_image, lightness_adjustment)
-
+        staged_image = filters.brightness(staged_image, lightness_adjustment)
+        print("Success 3")
         image_components.append(staged_image)
 
     output_image = filters.composite(image_components)
@@ -99,5 +76,6 @@ def apply_stencil(data, paths, resourcepack):
     if not os.path.exists(os.path.split(full_path_output)[0]):
         os.makedirs(os.path.split(full_path_output)[0])
 
-    os.remove(full_path_output)
+    if os.path.exists(full_path_output):
+        os.remove(full_path_output)
     output_image.save(full_path_output)
